failed attempt to transfer messages to Redux
