+++ static_src/containers/Layout.jsx
@@ -19,62 +19,8 @@ class Layout extends React.Component {
     chatId: 1,
   }
 
-  state = {
-    // chats: {
-    //   1: { title: 'Чат 1', messageList: [1] },
-    //   2: { title: 'Чат 2', messageList: [2] },
-    //   3: { title: 'Чат 3', messageList: [3] },
-    //   4: { title: 'Чат 4', messageList: [4] },
-    //   5: { title: 'Чат 5', messageList: [5] },
-    // },
-    messages: {
-      1: { text: 'Привет! я бот', sender: 'bot' },
-      2: { text: 'Здравствуйте! как дела?', sender: 'bot' },
-      3: { text: 'Привет! что нового?', sender: 'bot' },
-      4: { text: 'Здравствуйте! как у вас настроение?', sender: 'bot' },
-      5: { text: 'Привет! чё кислый такой?', sender: 'bot' },
-    },
-    input: '',
-  }
+  state = {}
 
-  // addChat = (title) => {
-  //   const { chats } = this.state
-  //   const chatId = Object.keys(chats).length + 1
-  //   this.setState({
-  //     chats: {
-  //       ...chats,
-  //       [chatId]: { title: title, messageList: [] },
-  //     },
-  //   })
-  // }
-
-  componentDidUpdate(prevProps, prevState) {
-    const { messages } = this.state
-    if (Object.keys(prevState.messages).length < Object.keys(messages).length
-      &&
-      Object.values(messages)[Object.values(messages).length - 1].sender
-      === 'me') {
-      const timeout = setTimeout(() =>
-        this.sendMessage('Не приставай ко мне, я робот!', 'bot'), 1000)
-    }
-  }
-
-  componentWillUnmount() {
-    clearTimeout(this.timeout)
-  }
-
-  sendMessage = (message, sender) => {
-    const { messages } = this.state
-    const { chatId } = this.props
-    const messageId = Object.keys(messages).length + 1
-    this.setState({
-      messages: {
-        ...messages,
-        [messageId]: { text: message, sender: sender },
-      },
-    })
-    this.props.sendMessage(messageId, message, sender, chatId)
-  }
 
   render() {
     return (
